Add files via upload
our study investigated the use of deep learning
models for the categorization of skin cancer, an important and
difficult issue in the discipline of dermatology. LeNet,
VGG19, InceptionV3, and DenseNet were four well-known
modes that we assessed for their ability to identify various
forms of skin cancer. Our results show that DenseNet is the
most precise and trustworthy model, with an excellent
accuracy rate of 86%. This demonstrates its outstanding
ability to correctly categorise cases of benign, basal cell
carcinoma, and melanoma. Although results from other
models, like VGG19 and InceptionV3, were similarly
encouraging, DenseNet’s performance highlights the model’s
potential for use in actual skin cancer diagnosis and early
detection. The success of DensNet in our work opens the door
for the creation of a reliable and effective automated skin
cancer diagnosis system in the future.